pyproject.toml
[project]
name = "context-haystack"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.9"
keywords = []
authors = [
  { name = "Alec Barber", email = "alec@context.ai" },
  { name = "Alex Gamble", email = "alex@context.ai" },
]
dependencies = [
  "haystack-ai",
  "context-python",
]

[project.urls]
Documentation = "https://docs.context.ai"
Issues = "https://github.com/contextco/context-py/issues"
Source = "https://github.com/contextco/context-py"
